Speed up log lookup in UI backend
We've run into scaling issues with the STDOUT/STDERR log view feature in the
Metaflow UI.  The current code path does an expensive join query to
determine the latest attempt, which ends up timing out.  However, the
relevant request is already supplying the attempt_id provided by the UI
from an earlier query, so there is no need to perform this verification.

Additional unrelated change: one of the existing tests was failing due
to the async duration not falling within the expected tolerance.  This
tolerance appeared to be specified incorrectly, using '1000' as a
relative tolerance.  I assume this was meant to be absolute tolerance,
as in 'this request should take (0 +- 1000 ms) to complete."  In my
development environment I needed 5000 to guarantee success with the
async nature of the tests.